Add files via upload
Image files to be shown on the ePaper screen.

world_map.bmp     Map of the World, but old version, I didn't like it.
WorldMap.bmp      Map of the World to show, over it, where is the ISS.
iss.bmp                   Small ISS image. 
space.bmp              Background to show the list of names of the people that is at the space.